Create Become a doctor from Russian Medical universities in low price 2021-22 Twinkle InstituteAB
Among the different developed nations all across the world, Russia is always being a highly educated and qualified country with a literacy rate of about 99%, also with a majority of its people residing there with a university degree holder. The high tradition of Russian medical educational institutions provides their students as well as international students travelling there with impressive studying and better living conditions as well. It is a heavenly beautiful place for every foreign student as every year Russia takes admission of thousands of medical students from across over 170 different countries in the world. Apart from this, the total cost of Study MBBS in Russia is very much affordable for an Indian family as compared to other developed countries like the US, UK, or Europe.

As we all know Russia is holding one of the largest land areas which give a chance from over 730 different universities in about 80 regions offering over more than 400 different fields of studying and more than 650 subjects across various managements. In addition to this, many of the major Medical college in Russia are running joint programs in association with other European universities which gives the benefit of accomplishing dual state degrees. A student pursuing Medicine in Russia gets the benefit of high-quality education with excellent sports facility also other co-curricular activities. After the study, over 60% of students in Russia are involved in sports activities and many universities provide excellent sports facilities to the students and get them plenty of opportunities to get into many of the top player’s teams with the best internationally renowned sports centres.

The facilities of healthcare and treatment in Russia is being provided by the government as well as by hospitals, clinics around. There are almost five thousand five hundred hospitals in Russia country. However, the most important thing to note is that the facilities of medical treatment provided in Russia are magnificent. Also, Russia has introduced some of the most specializations in the field of medicine, recently as they have successfully introduced the vaccine of COVID-19 and also has had many advanced achievements in different other fields.

Now Indian medical students who are dreaming of becoming a doctor and wish for studying MBBS abroad, especially MBBS in Russia are on the right page. Twinkle Institute will guide you in every best possible way to make your dream come true by becoming a doctor. So just click on our site for more information.

Most of the medical students of India after qualifying for the medical entrance test which is commonly called (NEET), have to go through a long procedure for getting a medical seat in the Indian government medical universities. As the availability is few and the demand for bulk of donation and high amount fees in private medical colleges leaves a student with very little hope. Taking these into consideration Russia is the appropriate place to pursue MBBS as it holds the Best Medical University in Russia without donation or high university fee structure. There are more than 15 best medical universities in Russia that are recognized by UNESCO, WHO, and MCI. Medical students from India who have studied in CBSE or ICSE boards can also study MBBS in Russia. The whole duration of medical courses in Russia is 6 years and besides which also includes one year of internship. The Medical college in Russia allow one year of internship where the cost of the fee is much cheaper during this time. Mostly, the MBBS course duration for medical education is almost the same all over the countries, but the only small differences of duration come due to additional language in different country’s universities. The normal duration is of 6 years which includes five years for the theory and one-year internship to study MBBS in Russia for foreign students who choose to study in an English taught program. However, if a residential student considers pursuing his MBBS education in the Russian language, then the duration is extended up to seven years that includes one extra year of language learning.

The admission criteria for studying MBBS in Russia in top medical universities is easy and is based on Online application and submission of your original documents to the selected medical university. If the applicant is eligible he/she will be informed to complete other verified formalities. This process starts in April and ends by October every year in most of the MBBS universities in Russia.