Create Medicine in Russia from Bashkir state medical university -Twinkle InstituteAB
BASHKIR STATE MEDICAL UNIVERSITY is one of the tops and best medical college in Russia. It is situated in the Ufa region in Russia. This university was established in the year 1932 and is ranked among the top 10 medical universities in Russia for providing the study of MBBS in Russia. This academy is registered in the Ministry of Education of the Russian federation. It is also recognized by MCI (Medical Council of India), Malaysia (MMC), the UK (GMC) and many other countries. Currently, the Bashkir state medical university offers degrees in several fields of medicine such as surgeons, oncologists, neurologists, dermatologists, morphologists, dentistry, pharmacy, paediatrics and the area of biochemists. Bashkir State Medical University has employed over 248 doctors in the department of medicine and 610 postgraduate for teaching. The university provides 85 departments for theoretical and medical education.

For all the undergraduate and postgraduate courses, Bashkir State Medical University has become the largest medical centre for providing MBBS education. It provides 56 scientific majors and doctoral courses in 8 scientific measures to the students. All these are accountable for defence and awarding academic degrees in 23 systems as well.

Overall we can conclude that this is among those Best Medical University In Russia where a student will be going to adapt to the modern lifestyle very easily. They have emanated liabilities, and the clinic for both medical and dental services is being provided over there. A student can easily lift them and set up a new peak in their career as well.

There are over 400 international medical students from 28 different countries are enrolled in this university every year. For Indian Students, this university is the best to be applied for, as this university demands low university fees for the MBBS study program.

There will be so many advantages and individually be going to get when they are coming for pursuing  MBBS Abroad in RUSSIA MEDICAL COLLEGE at Bashkir State Medical University. These are as follows:-

Fees Affordability:
When students are looking forward to keeping themselves with some affordable option then Bashkir State Medical University will hit at the top. A student will be able to complete their MBBS study within their budget as fees are quite cost-friendly.

Educational standards:
The Standard of education is high and settled up by university for students who are up to the mark. There will be no requirement for a student to compromise with their academic and non-academic activities. They will be going to have a clear mindset about both the particulars.

Hostel facilities:
Hostel facilities are also provided to the students within the university campus. There is no need for them to feel like they need to rent an apartment there. Until and unless they are agreeable with hostel facilities, they can continue with the same. In case they are looking forward to shifting to a private Apartment in Russia, then they need to complete all the mentioned guidelines the university has settled up.

Canteen facilities:
Usually, students face the trouble of getting Indian food when they are travelling for MBBS in Russia. The mess facilities are also there for them. The canteen has a separate menu for Indian medical students. There will be no chance for students to miss Indian food at all.

Internship program:
After completion of the MBBS study, the Bashkir state medical university will provide the internship program to every student studying at Bashkir state medical university. The internship program will let them understand and apply what they have learned in the previous year so that they can easily move ahead to practice medicine, and no trouble will be there for students in any case.

Security checks:
All in all the security guards, as well as there, are 24×7 CCTV checks on every floor and doors. In case there is any problem regarding security arising with them they can simply report it to the higher authorities present there. There will be going to take strict actions without creating any delay in this university.

The duration of MBBS study in Russia at Bashkir State Medical University is almost 6 years. It will be going to a five-year academic session and about one year of the internship program.

Bashkir State Medical University is one of the best choices for the medical aspirants for those who want to pursue Medicine in Russia. The major attractive elements are its low-cost accommodation and low-cost tuition fees structure. MBBS in Bashkir state medical university suits every medical student. MBBS degree at Bashkir State Medical University costs around 20 lakhs for the whole course. This cost is much cheaper as compared to other medical universities. This university offers a 6-year MBBS degree program to every international student. Final year training is well-organized in reputed government hospitals. On the other hand, these medical university hostels can accommodate 2000 students in one year. Each hostel room occupies an attached kitchen and washroom. Internet facility is also being provided to the students free of cost.  Indian food is available every day in the hostel canteen to the students. Indians are treated with love and respect by the Russians. Bashkir State Medical University often assembles cultural events for students every year. Graduates from Bashkir State Medical University can do practice medicine in any part of the world.