Create MBBS Admission Criteria in Russia for Indian Students-Twinkle InstituteAB
The whole process of getting admission to BEST MEDICAL UNIVERSITY IN RUSSIA for MBBS is very simple. Russian Medical Universities do not require Indian students to clear any entrance exam for admission to study MBBS Course. Admissions are based on the marks scored in the 12th board examination (intermediate). Candidates applying for MBBS admission in Russia must also fulfill the guidelines of Russian medical universities. Also, NEET-UG eligibility criteria set by the Medical Council of India.

The application process for admission to MBBS in Russia will likely start from May every year. The application form filing period is from May to September. All documents must be submitted by September 15 to proceed with admission formalities. We advise students to apply for early admission because there is a large number of applicants and seats are limited. MBBS Abroad admissions in the Russian medical universities are given strictly on a first come first served basis and depend on availability. Therefore, there is an added benefit to those candidates who apply early.

The academic year in Russia medical college has two semesters. The first semester of RUSSIAN MBBS COLLEGE starts in mid-September and ends in February first week and if we talk about the second semester it starts in the middle of February and ends in June and July. At the end of each semester, students are passing their examinations. After this, they have a vacation: 2 weeks in winter (from 25th of January to 9th of February), 2 months in summer (from mid-June to mid of August).

Russia is the perfect destination for Indian medical students to study MBBS mainly because there is no competitive exam or huge donation required for taking admission to the Russian universities. As we have mentioned above, the admission procedure is easy and simple. The course fees are low and subsidized. The MBBS degree is recognized worldwide provided by Top Medical Universities in Russia as it is approved by all the top world’s medical councils. Healthy food and accommodation are provided, also, there are Indian canteens available in most of the MBBS colleges in Russia. Universities provide MCI coaching in the Russian Medical Universities and the results are good every year. Students that have graduated from these universities are working in leading hospitals and bodies all over the world. The MBBS graduates are not only working in India but in developed countries like Australia, Canada, the USA, UK, and other major countries across the world. Also, there is a very less percentage of rejecting VISA. All the universities are government bodies in Russia.

Eligibility Criteria for MBBS in Russia

The eligibility criteria of the students who want to pursue MBBS in Russia are:-

• The candidate should be of the age 17-30. They should complete the age of 17 years on or before 31st December of the year of application.

• The candidate must secure 60% marks in their 12th grade from the Science stream (Physics, Chemistry, and Biology) for applying for MBBS in Russia.

• The candidates of the SC/ST/OBC categories need to score a minimum of 50% marks in their 12th standard.

• As per the latest notification by NMC, Indian medical candidates who plan to study in Foreign Medical Universities need to qualify NEET for practicing in their country.

For MBBS admission in Russia for Indian students fee structure is also not very high barring except few medical universities. The MBBS tuition fee is really low cost and in the budget for most of the Indian aspirants which makes them choose Russia as a destination for MBBS abroad.

Instead of running behind the lowest MBBS fees package in Russia which can be around 2500 USD per annum as the tuition fees, but we will always guide our students to get admissions in the good  Russian medical universities only. Eligibility criteria for pursuing MBBS in Russia is also very basic which gives students chance to study medicine without clearing any competitive exams.

Russia is a developed nation with good infrastructure and every facility. Medical schools in Russia have got high technology in education and medication. The MBBS degrees earned from the best medical Universities in Russia are approved and recognized worldwide. The degree is also recognized by the WHO, FAIMER, ECFMG, and more government medical bodies. This degree recognition makes the students eligible to appear for the licensing exams to practice medicine in various countries all over the globe. The Russian medical universities and colleges provide Medicine courses in both English and Russian language, you can opt for any. Though they even had bilingual MBBS programs in which they taught students in English for the first 3 years and then in Russia for the latter 3 years of the course.